api/requirements.txt
fastapi
uvicorn
transformers
diffusers
torch
opencv-python
Pillow
accelerate
python-multipart
numpy
scikit-image
xformers
safetensors
controlnet-aux
